EDITSKS-5515: impl long to string converter (separate from long? to string)
