+++ Cassandra.DistributedTaskQueue.Monitoring/Json/LongToStringConverter.cs
@@ -1,36 +1,23 @@
-﻿#nullable enable
-
-using System;
+﻿using System;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 
 namespace SkbKontur.Cassandra.DistributedTaskQueue.Monitoring.Json;
 
 /// <summary>
-///     Конвертер, которые сериализует long-и в строки вместо чисел,
-///     ибо JSON.parse() на длинных long-ах теряет последние разряды
+///     Конвертор для сериализации long в строку
+///     note: JSON.parse() при сериализации длинных чисел теряет последние разряды
 /// </summary>
-public class LongToStringConverter : JsonConverter<long?>
+public class LongToStringConverter : JsonConverter<long>
 {
-    public override long? Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
+    public override long Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
     {
-        var value = JsonSerializer.Deserialize<string?>(ref reader);
-        if (value == null)
-        {
-            return null;
-        }
+        var value = JsonSerializer.Deserialize<string>(ref reader);
         return long.Parse(value);
     }
 
-    public override void Write(Utf8JsonWriter writer, long? value, JsonSerializerOptions options)
+    public override void Write(Utf8JsonWriter writer, long value, JsonSerializerOptions options)
     {
-        if (value == null)
-        {
-            writer.WriteNullValue();
-        }
-        else
-        {
-            writer.WriteStringValue(value.ToString());
-        }
+        writer.WriteStringValue(value.ToString());
     }
 }
