+++ version.json
@@ -1,5 +1,5 @@
 {
-  "version": "3.4-prerelease3",
+  "version": "3.4",
   "assemblyVersion": {
     "precision": "build"
   },
